+++ src/tree.js
@@ -1,6 +1,6 @@
 import _ from 'lodash/fp'
-import { push } from './array'
 import { findIndexed } from './conversion'
+import { push, dotEncoder, slashEncoder } from './array'
 
 export let isTraversable = x => _.isArray(x) || _.isPlainObject(x)
 export let traverse = x => isTraversable(x) && !_.isEmpty(x) && x
@@ -18,6 +18,13 @@ export let walk = (next = traverse) => (
   ) ||
   post(tree, index, parents, parentIndexes)
 
+export let transformTree = (next = traverse) =>
+  _.curry((f, x) => {
+    let result = _.cloneDeep(x)
+    walk(next)(f)(result)
+    return result
+  })
+
 export let reduceTree = (next = traverse) =>
   _.curry((f, result, tree) => {
     walk(next)((...x) => {
@@ -33,10 +40,53 @@ export let treeToArray = (next = traverse) => treeToArrayBy(next)(x => x)
 export let leaves = (next = traverse) =>
   _.flow(treeToArray(next), _.reject(next))
 
-export let tree = (next = traverse) => ({
+export let treeLookup = (next = traverse, buildIteratee = _.identity) => (
+  path,
+  tree
+) =>
+  _.reduce((tree, path) => _.find(buildIteratee(path), next(tree)), tree, path)
+
+export let keyTreeByWith = (next = traverse) =>
+  _.curry((transformer, groupIteratee, x) =>
+    _.flow(
+      treeToArrayBy(next)(_.iteratee(groupIteratee)),
+      _.uniq,
+      _.keyBy(_.identity),
+      _.mapValues(group =>
+        transformTree(next)(node => {
+          let matches = _.iteratee(groupIteratee)(node) === group
+          transformer(node, matches, group)
+        }, x)
+      )
+    )(x)
+  )
+
+// Flat Tree
+export let treeKeys = (x, i, xs, is) => [i, ...is]
+export let treeValues = (x, i, xs) => [x, ...xs]
+export let treePath = (build = treeKeys, encoder = dotEncoder) => (...args) =>
+  (encoder.encode || encoder)(build(...args).reverse())
+export let propTreePath = prop =>
+  treePath(_.flow(treeValues, _.map(prop)), slashEncoder)
+
+export let flattenTree = (next = traverse) => (buildPath = treePath()) =>
+  reduceTree(next)(
+    (result, node, ...x) => _.set([buildPath(node, ...x)], node, result),
+    {}
+  )
+
+export let flatLeaves = (next = traverse) => _.reject(next)
+
+export let tree = (next = traverse, buildIteratee = _.identity) => ({
   walk: walk(next),
+  transform: transformTree(next),
   reduce: reduceTree(next),
   toArrayBy: treeToArrayBy(next),
   toArray: treeToArray(next),
   leaves: leaves(next),
+  lookup: treeLookup(next, buildIteratee),
+  keyByWith: keyTreeByWith(next),
+  traverse: next,
+  flatten: flattenTree(next),
+  flatLeaves: flatLeaves(next),
 })